Split CLI diagnostic rendering
